corban-msc-kafka/requirements.txt
# --- App dependencies ---
fastapi
uvicorn
aiohttp
aiokafka
pydantic

# --- OpenTelemetry Auto Instrumentation ---
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-asyncio
opentelemetry-instrumentation-aiohttp-client
opentelemetry-instrumentation-aiokafka
opentelemetry-instrumentation-requests

